LayoutTests/platform/mac/fast/speechsynthesis/speech-synthesis-speak-expected.txt

This tests that the basic mechanisms of speaking text work, including sending the job and receving the callback.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS speechSynthesis.speaking is false
PASS event.elapsedTime > 0 is true
PASS speechSynthesis.speaking is true
PASS event.elapsedTime > 0 is true
PASS speechSynthesis.speaking is false
PASS successfullyParsed is true

TEST COMPLETE
